Refactor code smells in items.cpp and about_window.cpp
- Extracted `GamePanel`, `TetrisPanel`, and `SnakePanel` from `source/ui/about_window.cpp` into separate files in `source/ui/minigames/`.
- Refactored `ItemDatabase::loadFromOtbVer1` in `source/game/items.cpp` by extracting attribute parsing logic into `parseOtbAttrVer1`.
- Refactored `ItemDatabase::loadItemFromGameXml` in `source/game/items.cpp` by extracting attribute parsing logic into `parseGameXmlAttribute`.
- Updated `source/CMakeLists.txt` to include the new minigame source files.
